backend/requirements/base.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic[email]>=2.5.0
transformers>=4.35.2
torch>=2.1.0
Pillow>=10.1.0
boto3>=1.28.64
python-magic>=0.4.27
pytest>=7.4.3
httpx>=0.25.1
psycopg2-binary>=2.9.9
python-multipart>=0.0.6